Restructure cc and cf
